feat: Export Auth type
