+++ .changeset/eleven-mirrors-fry.md
@@ -1,0 +1,5 @@
+---
+'@bnb-chain/greenfield-js-sdk': patch
+---
+
+feat: Export Auth type
